# wp-content/plugins/gym-core/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gym Core REST API",
    "description": "REST API for Haanpaa Martial Arts gym management. All endpoints use the gym/v1 namespace and return a standard envelope: { success: true, data: <payload>, meta?: { pagination? } }.",
    "version": "2.1.0",
    "contact": {
      "name": "Andrew Wikel"
    }
  },
  "servers": [
    {
      "url": "/wp-json/gym/v1",
      "description": "WordPress REST API (gym/v1 namespace)"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WP-Nonce",
        "description": "WordPress REST nonce (cookie authentication)"
      },
      "applicationPassword": {
        "type": "http",
        "scheme": "basic",
        "description": "WordPress Application Password"
      }
    },
...
  },
  "paths": {
...
    "/briefings/today": {
      "get": {
        "summary": "Get briefings for all today's classes",
        "tags": ["Briefings"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "parameters": [
          { "name": "location", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "All briefings for today, sorted by start time" }
        }
      }
    },
    "/announcements": {
      "get": {
        "summary": "List active announcements",
        "tags": ["Briefings"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "parameters": [
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "program", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 10, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Paginated active announcements" }
        }
      },
      "post": {
        "summary": "Create an announcement",
        "tags": ["Briefings"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "type": { "type": "string", "enum": ["global", "location", "program"], "default": "global" },
                  "target_location": { "type": "string" },
                  "target_program": { "type": "string" },
                  "start_date": { "type": "string", "format": "date" },
                  "end_date": { "type": "string", "format": "date" },
                  "pinned": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Announcement created" },
          "403": { "description": "Requires gym_manage_announcements" }
        }
      }
    },
...
  },
...
}